Move logic from OrderRepository to OrderService.
Moved some logic which was in OrderRepository to OrderService where it made more sense. Corrected the type of OrderRepository to be an interface instead of class. Optimized imports across the project. Changed Hibernate's schema generation back to 'create-drop' which I had changed previously to 'update' because of a Hibernate error upon application start having to do with trying to remove a foreign key on a table that didn't exist. That error now seems to be gone, possibly due to the correctly implemented Customer Order relationship.